refactor(iota-types): isolate wasm-incompatible code into sibling files
Move the wasm-specific code out of the shared source files so the native
files stay byte-identical to develop and the gating lives at the module
boundary instead of scattered inline:

- metrics: revert metrics.rs to its original prometheus-backed form and
  put the no-op stubs in wasm_metrics.rs; lib.rs selects the module per
  target. The wasm build still needs the metric types (execution +
  signature_verification), so the impl is swapped, not dropped.
- messages_consensus: the DKG subset is the only wasm-incompatible part
  (pulls fastcrypto-tbls) and nothing on the wasm path uses it, so it
  moves to messages_consensus_dkg.rs and is cfg'd out entirely on wasm.
- multiaddr: wire wasm_multiaddr.rs via #[path] for consistency.
