+++ crates/iota-types/src/lib.rs
@@ -13,11 +13,8 @@ use base_types::{IotaAddress, ObjectID, SequenceNumber};
 #[cfg(not(target_arch = "wasm32"))]
 pub use iota_network_stack::multiaddr;
 #[cfg(target_arch = "wasm32")]
-pub mod multiaddr {
-    pub use crate::wasm_multiaddr::Multiaddr;
-}
-#[cfg(target_arch = "wasm32")]
-mod wasm_multiaddr;
+#[path = "wasm_multiaddr.rs"]
+pub mod multiaddr;
 pub use iota_sdk_types as sdk_types;
 use iota_sdk_types::{StructTag, TypeTag};
 use move_binary_format::{
@@ -78,6 +75,10 @@ pub mod messages_checkpoint;
 pub mod messages_consensus;
 pub mod messages_grpc;
 pub mod messages_safe_client;
+#[cfg(not(target_arch = "wasm32"))]
+pub mod metrics;
+#[cfg(target_arch = "wasm32")]
+#[path = "wasm_metrics.rs"]
 pub mod metrics;
 pub mod mock_checkpoint_builder;
 pub mod move_authenticator;
